requirements.txt
apipkg==1.5
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
Click==7.0
colorama==0.4.1
contextlib2==0.5.5
execnet==1.5.0
Flask==1.0.2
importlib-metadata==0.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mock==2.0.0
more-itertools==4.3.0
neo4j-driver==1.6.2
neotime==1.0.0
numpy==1.16.0
opencv-python==4.0.0.21
pandas==0.23.4
path.py==11.5.0
pbr==5.1.1
Pillow==5.4.1
pluggy==0.8.0
prompt-toolkit==1.0.15
py==1.7.0
py2neo==4.1.3
Pygments==2.3.0
pyparsing==2.3.0
pytest==3.10.1
pytest-shutil==1.4.0
python-dateutil==2.7.5
pytz==2018.7
scikit-learn==0.20.2
scipy==1.2.0
six==1.11.0
sklearn==0.0
termcolor==1.1.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
zipp==0.3.3
